Exception handling - delete

+++ src/main/java/github/wescaetano/first_spring_boot_project/services/UserService.java
@@ -2,8 +2,11 @@ package github.wescaetano.first_spring_boot_project.services;
 
 import github.wescaetano.first_spring_boot_project.entities.User;
 import github.wescaetano.first_spring_boot_project.repositories.UserRepository;
+import github.wescaetano.first_spring_boot_project.services.exceptions.DatabaseException;
 import github.wescaetano.first_spring_boot_project.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,8 +31,16 @@ public class UserService {
         return repository.save(obj);
     }
 
-    public void delete(Long id){
-        repository.deleteById(id);
+    public void delete(Long id) {
+        Optional<User> user = repository.findById(id);
+        if (!user.isPresent()) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
